pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "extract-msg>=0.54.1",
    "pandas>=2.3.0",
    "pathlib>=1.0.1",
    "pdfplumber>=0.11.7",
    "pillow>=11.2.1",
    "plotly>=6.2.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "spacy>=3.8.7",
    "streamlit>=1.46.1",
]
